+++ poc/cpat-client/src/state-management/location/services.js
@@ -1,16 +1,11 @@
+import { handlers } from '../../state-management/helpers/http-response-handler';
+
 export const locationService = {
     getSingle,
     getList,
     insert,
     update,
     remove
-};
-
-/**
- * Holds configuration info for interacting with any attached APIs.
- */
-const config = {
-    apiUrl: process.env.REACT_APP_API_URL || 'https://localhost:5001'
 };
 
 /**
@@ -23,7 +18,7 @@ function getSingle(id) {
         headers: { 'Content-Type': 'application/json' },
     };
 
-    return fetch(`${config.apiUrl}/table/${id}`, requestOptions).then(handleResponse);
+    return fetch(`${handlers.config.apiUrl}/table/${id}`, requestOptions).then(handlers.handleHttpResponse);
 };
 
 /**
@@ -36,7 +31,7 @@ function getList(idList) {
         headers: { 'Content-Type': 'application/json' },
     };
 
-    return fetch(`${config.apiUrl}/table/${idList}`, requestOptions).then(handleResponse);
+    return fetch(`${handlers.config.apiUrl}/table/${idList}`, requestOptions).then(handlers.handleHttpResponse);
 };
 
 /**
@@ -50,7 +45,7 @@ function insert(locationDoc) {
         body: JSON.stringify(locationDoc)
     };
 
-    return fetch(`${config.apiUrl}/table`, requestOptions).then(handleResponse);
+    return fetch(`${handlers.config.apiUrl}/table`, requestOptions).then(handlers.handleHttpResponse);
 };
 
 /**
@@ -64,7 +59,7 @@ function update(locationDoc) {
         body: JSON.stringify(locationDoc)
     };
 
-    return fetch(`${config.apiUrl}/table`, requestOptions).then(handleResponse);
+    return fetch(`${handlers.config.apiUrl}/table`, requestOptions).then(handlers.handleHttpResponse);
 };
 
 /**
@@ -78,28 +73,5 @@ function remove(id) {
         body: JSON.stringify(id)
     };
 
-    return fetch(`${config.apiUrl}/table`, requestOptions).then(handleResponse);
+    return fetch(`${handlers.config.apiUrl}/table`, requestOptions).then(handlers.handleHttpResponse);
 };
-
-
-/**
- * 
- * @param {*} response 
- */
-function handleResponse(response) {
-    return response.text().then(text => {
-        const data = text && JSON.parse(text);
-        if (!response.ok) {
-            if (response.status === 401) {
-                // auto logout if 401 response returned from api
-                // authenticationService.logout();
-                Location.reload(true);
-            }
-
-            const error = (data && data.message) || response.statusText;
-            return Promise.reject(error);
-        }
-
-        return data;
-    });
-}
